// web/src/openapi/model/programme.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DOCUMENTATION
 * OpenAPI spec version: 1.0.0
 */
import type { Event } from './event';
import type { ProgrammeCover } from './programmeCover';
import type { ProgrammeCreatedBy } from './programmeCreatedBy';
import type { ProgrammeUniverse } from './programmeUniverse';
import type { ProgrammeUpdatedBy } from './programmeUpdatedBy';

export interface Programme {
  id?: string | number;
  documentId?: string;
  title?: string;
  cover: ProgrammeCover;
  description?: string;
  universe?: ProgrammeUniverse;
  slug?: string;
  events?: Event[];
  summary?: string;
  createdAt?: string;
  updatedAt?: string;
  publishedAt?: string;
  createdBy?: ProgrammeCreatedBy;
  updatedBy?: ProgrammeUpdatedBy;
  locale?: string;
  localizations?: Programme[];
}